+++ app/data_models/questionnaire_store.py
@@ -3,7 +3,7 @@ from types import MappingProxyType
 from app.data_models.answer_store import AnswerStore
 from app.data_models.list_store import ListStore
 from app.data_models.progress_store import ProgressStore
-from app.questionnaire.routing.utils import parse_iso_8601_datetime
+from app.questionnaire.rules.utils import parse_iso_8601_datetime
 from app.utilities.json import json_dumps, json_loads
